move files to module, add testing infrastructure
